added tags for notes (remains to add the logic in the archive page)

+++ src/Controller/NotesController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\Categories;
 use App\Entity\Notes;
 use App\Entity\Users;
 use App\Form\NoteFormType;
@@ -10,6 +11,8 @@ use phpDocumentor\Reflection\Types\Integer;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -96,16 +99,84 @@ class NotesController extends AbstractController
     {
         $params = json_decode($request->getContent(), true);
         foreach ($params["selectedNotes"] as $noteId) {
-           $note = $entityManager->getRepository(Notes::class)->find($noteId);
-           $note->setArchived(true);
-           $entityManager->persist($note);
+            $note = $entityManager->getRepository(Notes::class)->find($noteId);
+            $userID = $entityManager->getRepository(Users::class)->find($this->getUser()->getUserIdentifier());
+            if ($note->getIdUser() == $userID) {
+               $note->setArchived(true);
+               $entityManager->persist($note);
+               $entityManager->flush();
+           }
+        }
+        return new Response('success');
+    }
+
+    #[Route('/unarchiveNotes', name: 'unarchive_notes')]
+    public function unarchiveNotes(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $params = json_decode($request->getContent(), true);
+        foreach ($params["selectedNotes"] as $noteId) {
+            $note = $entityManager->getRepository(Notes::class)->find($noteId);
+            $userID = $entityManager->getRepository(Users::class)->find($this->getUser()->getUserIdentifier());
+            if ($note->getIdUser() == $userID) {
+                $note->setArchived(false);
+                $entityManager->persist($note);
+                $entityManager->flush();
+            }
+        }
+        return new Response('success');
+    }
+
+    #[Route('/TagNotes', name: 'tag_notes')]
+    public function tagNotes(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $params = json_decode($request->getContent(), true);
+
+        $tag = $entityManager->getRepository(Categories::class)->findOneBy(array('name' =>$params['selectedTag']));
+        if ($tag == null) {
+            $tag = new Categories();
+
+            $tag->setName(str_replace(' ', '', strtolower($params['selectedTag'])));
+            $entityManager->persist($tag);
+            $entityManager->flush();
         }
 
-        $entityManager->flush();
+        $tag = $entityManager->getRepository(Categories::class)->findOneBy(array('name' =>(strtolower($params['selectedTag'])) ));
+        foreach ($params["selectedNotes"] as $noteId) {
 
+            $note = $entityManager->getRepository(Notes::class)->find($noteId);
+            if ($note == null) {
+                return new Response('error');
+            }
+            $userID = $entityManager->getRepository(Users::class)->find($this->getUser()->getUserIdentifier());
+            if ($note->getIdUser() == $userID) {
+                $note->addIdCategory($tag);
+                $entityManager->persist($note);
+                $entityManager->flush();
+            }
+        }
 
         return new Response('success');
     }
+
+    #[Route('/removeTagFromNote/{noteId}&{tagId}', name: 'remove_tag_note')]
+    public function removeTagFromNote(int $noteId, int $tagId, Request $request, EntityManagerInterface $entityManager): Response
+    {
+
+        $tag = $entityManager->getRepository(Categories::class)->find($tagId);
+        $note = $entityManager->getRepository(Notes::class)->find($noteId);
+        if ($tag == null || $note == null) {
+            return new Response('error');
+        }
+
+        $note->removeIdCategory($tag);
+        $entityManager->persist($note);
+        $entityManager->flush();
+        return $this->redirectToRoute('dashboard');
+    }
+
+
+
+
 
     #[Route('/deleteNotes', name: 'delete_notes')]
     public function deleteNotes(Request $request, EntityManagerInterface $entityManager): Response
@@ -113,16 +184,12 @@ class NotesController extends AbstractController
         $params = json_decode($request->getContent(), true);
         foreach ($params["selectedNotes"] as $noteId) {
             $note = $entityManager->getRepository(Notes::class)->find($noteId);
-            $note->setArchived(true);
-            $entityManager->remove($note);
+            $userID = $entityManager->getRepository(Users::class)->find($this->getUser()->getUserIdentifier());
+            if ($note->getIdUser() == $userID) {
+                $entityManager->remove($note);
+                $entityManager->flush();
+            }
         }
-
-        $entityManager->flush();
-
-
         return new Response('success');
     }
-
-
-
 }
